Moved string functions to lib file
