+++ runtime/src/runtime.js
@@ -1,5 +1,7 @@
 import { heapAdjust } from './core/heap';
 
+import * as libStrings from './lib/strings';
+
 let grainInitialized = false;
 let grainModule;
 let grainDOMRefs = [];
@@ -107,39 +109,6 @@ function printNumber(n) {
   return n;
 }
 
-
-
-function stringAppend(s1, s2) {
-  assertString(s1);
-  assertString(s2);
-  s1 = grainToJSVal(s1);
-  s2 = grainToJSVal(s2);
-  let appended = s1.concat(s2);
-  let ret = JSToGrainVal(appended);
-  return ret;
-}
-
-function stringLength(s) {
-  assertString(s);
-  return JSToGrainVal(grainToJSVal(s).length);
-}
-
-function stringSlice(s, from, to) {
-  assertString(s);
-  assertNumber(from);
-  assertNumber(to);
-  s = grainToJSVal(s);
-  if (from < 0) {
-    throwGrainError(GRAIN_ERR_NOT_NONNEG, from);
-  } else if ((to >> 1) > s.length) {
-    throw new GrainError(-1, `Index ${to >> 1} greater than string length (${s.length})`);
-  }
-  from = grainToJSVal(from);
-  to = grainToJSVal(to);
-  let ret = s.slice(from, to);
-  return JSToGrainVal(ret);
-}
-
 const importObj = {
   console: {
     log: printNumber,
@@ -155,9 +124,7 @@ const importObj = {
     print: printNumber,
     equal: grainEqual,
     toString: (n => JSToGrainVal(grainToString(n))),
-    stringAppend: stringAppend,
-    stringLength: stringLength,
-    stringSlice: stringSlice,
+    ...libStrings,
     DOMQuery: grainDOMQuery,
     DOMSetText: grainDOMElemSetText,
     DOMDangerouslySetInnerHTML: grainDOMDangerouslySetInnerHTML,
